+++ RealmSwift/List.swift
@@ -192,6 +192,23 @@ public final class List<T: Object>: ListBase {
     /**
      Returns a `Results` containing the objects in the list, but sorted.
 
+     Objects are sorted based on the values of the given key path. For example, to sort a list of `Student`s from
+     youngest to oldest based on their `age` property, you might call
+     `students.sorted(byKeyPath: "age", ascending: true)`.
+
+     - warning: Lists may only be sorted by properties of boolean, `Date`, `NSDate`, single and double-precision
+                floating point, integer, and string types.
+
+     - parameter keyPath:  The key path to sort by.
+     - parameter ascending: The direction to sort in.
+     */
+    public func sorted(byKeyPath keyPath: String, ascending: Bool = true) -> Results<T> {
+        return sorted(by: [SortDescriptor(keyPath: keyPath, ascending: ascending)])
+    }
+
+    /**
+     Returns a `Results` containing the objects in the list, but sorted.
+
      Objects are sorted based on the values of the given property. For example, to sort a list of `Student`s from
      youngest to oldest based on their `age` property, you might call
      `students.sorted(byProperty: "age", ascending: true)`.
@@ -202,8 +219,9 @@ public final class List<T: Object>: ListBase {
      - parameter property:  The name of the property to sort by.
      - parameter ascending: The direction to sort in.
      */
+    @available(*, deprecated, renamed: "sorted(byKeyPath:ascending:)")
     public func sorted(byProperty property: String, ascending: Bool = true) -> Results<T> {
-        return sorted(by: [SortDescriptor(property: property, ascending: ascending)])
+        return sorted(byKeyPath: property, ascending: ascending)
     }
 
     /**
@@ -212,7 +230,7 @@ public final class List<T: Object>: ListBase {
      - warning: Lists may only be sorted by properties of boolean, `Date`, `NSDate`, single and double-precision
                 floating point, integer, and string types.
 
-     - see: `sorted(byProperty:ascending:)`
+     - see: `sorted(byKeyPath:ascending:)`
     */
     public func sorted<S: Sequence>(by sortDescriptors: S) -> Results<T> where S.Iterator.Element == SortDescriptor {
         return Results<T>(_rlmArray.sortedResults(using: sortDescriptors.map { $0.rlmSortDescriptorValue }))
@@ -517,7 +535,7 @@ extension List {
     @available(*, unavailable, renamed: "index(matching:_:)")
     public func index(of predicateFormat: String, _ args: Any...) -> Int? { fatalError() }
 
-    @available(*, unavailable, renamed: "sorted(byProperty:ascending:)")
+    @available(*, unavailable, renamed: "sorted(byKeyPath:ascending:)")
     public func sorted(_ property: String, ascending: Bool = true) -> Results<T> { fatalError() }
 
     @available(*, unavailable, renamed: "sorted(by:)")
@@ -712,17 +730,17 @@ public final class List<T: Object>: ListBase {
     /**
      Returns a `Results` containing the objects in the list, but sorted.
 
-     Objects are sorted based on the values of the given property. For example, to sort a list of `Student`s from
+     Objects are sorted based on the values of the given key path. For example, to sort a list of `Student`s from
      youngest to oldest based on their `age` property, you might call `students.sorted("age", ascending: true)`.
 
      - warning: Lists may only be sorted by properties of boolean, `NSDate`, single and double-precision floating point,
                 integer, and string types.
 
-     - parameter property:  The name of the property to sort by.
+     - parameter keyPath:  The key path to sort by.
      - parameter ascending: The direction to sort in.
      */
-    public func sorted(property: String, ascending: Bool = true) -> Results<T> {
-        return sorted([SortDescriptor(property: property, ascending: ascending)])
+    public func sorted(keyPath: String, ascending: Bool = true) -> Results<T> {
+        return sorted([SortDescriptor(keyPath: keyPath, ascending: ascending)])
     }
 
     /**