+++ RealmSwift/Tests/RealmCollectionTypeTests.swift
@@ -311,15 +311,19 @@ class RealmCollectionTypeTests: TestCase {
         guard let collection = collection else {
             fatalError("Test precondition failed")
         }
-        var sorted = collection.sorted(byProperty: "stringCol", ascending: true)
+        var sorted = collection.sorted(byKeyPath: "stringCol", ascending: true)
         XCTAssertEqual("1", sorted[0].stringCol)
         XCTAssertEqual("2", sorted[1].stringCol)
 
-        sorted = collection.sorted(byProperty: "stringCol", ascending: false)
+        sorted = collection.sorted(byKeyPath: "stringCol", ascending: false)
         XCTAssertEqual("2", sorted[0].stringCol)
         XCTAssertEqual("1", sorted[1].stringCol)
 
-        assertThrows(collection.sorted(byProperty: "noSuchCol", ascending: true), named: "Invalid sort property")
+        sorted = collection.sorted(byKeyPath: "linkCol.id", ascending: true)
+        XCTAssertEqual("1", sorted[0].stringCol)
+        XCTAssertEqual("2", sorted[1].stringCol)
+
+        assertThrows(collection.sorted(byKeyPath: "noSuchCol", ascending: true), named: "Invalid property name")
     }
 
     func testSortWithDescriptor() {
@@ -329,19 +333,19 @@ class RealmCollectionTypeTests: TestCase {
         XCTAssertEqual(collection[0], notActuallySorted[0])
         XCTAssertEqual(collection[1], notActuallySorted[1])
 
-        var sorted = collection.sorted(by: [SortDescriptor(property: "intCol", ascending: true)])
+        var sorted = collection.sorted(by: [SortDescriptor(keyPath: "intCol", ascending: true)])
         XCTAssertEqual(1, sorted[0].intCol)
         XCTAssertEqual(2, sorted[1].intCol)
 
-        sorted = collection.sorted(by: [SortDescriptor(property: "doubleCol", ascending: false),
-            SortDescriptor(property: "intCol", ascending: false)])
+        sorted = collection.sorted(by: [SortDescriptor(keyPath: "doubleCol", ascending: false),
+            SortDescriptor(keyPath: "intCol", ascending: false)])
         XCTAssertEqual(2.22, sorted[0].doubleCol)
         XCTAssertEqual(3, sorted[0].intCol)
         XCTAssertEqual(2.22, sorted[1].doubleCol)
         XCTAssertEqual(2, sorted[1].intCol)
         XCTAssertEqual(1.11, sorted[2].doubleCol)
 
-        assertThrows(collection.sorted(by: [SortDescriptor(property: "noSuchCol")]), named: "Invalid sort property")
+        assertThrows(collection.sorted(by: [SortDescriptor(keyPath: "noSuchCol")]), named: "Invalid property name")
     }
 
     func testMin() {
@@ -843,9 +847,9 @@ class ListStandaloneRealmCollectionTypeTests: ListRealmCollectionTypeTests {
 
     override func testSortWithDescriptor() {
         let collection = getAggregateableCollection()
-        assertThrows(collection.sorted(by: [SortDescriptor(property: "intCol", ascending: true)]))
-        assertThrows(collection.sorted(by: [SortDescriptor(property: "doubleCol", ascending: false),
-            SortDescriptor(property: "intCol", ascending: false)]))
+        assertThrows(collection.sorted(by: [SortDescriptor(keyPath: "intCol", ascending: true)]))
+        assertThrows(collection.sorted(by: [SortDescriptor(keyPath: "doubleCol", ascending: false),
+            SortDescriptor(keyPath: "intCol", ascending: false)]))
     }
 
     override func testFastEnumerationWithMutation() {
@@ -872,8 +876,8 @@ class ListStandaloneRealmCollectionTypeTests: ListRealmCollectionTypeTests {
         guard let collection = collection else {
             fatalError("Test precondition failed")
         }
-        assertThrows(collection.sorted(byProperty: "stringCol", ascending: true))
-        assertThrows(collection.sorted(byProperty: "noSuchCol", ascending: true))
+        assertThrows(collection.sorted(byKeyPath: "stringCol", ascending: true))
+        assertThrows(collection.sorted(byKeyPath: "noSuchCol", ascending: true))
     }
 
     override func testFilterFormat() {
@@ -1365,7 +1369,11 @@ class RealmCollectionTypeTests: TestCase {
         XCTAssertEqual("2", sorted[0].stringCol)
         XCTAssertEqual("1", sorted[1].stringCol)
 
-        assertThrows(self.collection.sorted("noSuchCol", ascending: true), named: "Invalid sort property")
+        sorted = collection.sorted("linkCol.id", ascending: true)
+        XCTAssertEqual("1", sorted[0].stringCol)
+        XCTAssertEqual("2", sorted[1].stringCol)
+
+        assertThrows(self.collection.sorted("noSuchCol", ascending: true), named: "Invalid property name")
     }
 
     func testSortWithDescriptor() {
@@ -1375,19 +1383,19 @@ class RealmCollectionTypeTests: TestCase {
         XCTAssertEqual(collection[0], notActuallySorted[0])
         XCTAssertEqual(collection[1], notActuallySorted[1])
 
-        var sorted = collection.sorted([SortDescriptor(property: "intCol", ascending: true)])
+        var sorted = collection.sorted([SortDescriptor(keyPath: "intCol", ascending: true)])
         XCTAssertEqual(1, sorted[0].intCol)
         XCTAssertEqual(2, sorted[1].intCol)
 
-        sorted = collection.sorted([SortDescriptor(property: "doubleCol", ascending: false),
-            SortDescriptor(property: "intCol", ascending: false)])
+        sorted = collection.sorted([SortDescriptor(keyPath: "doubleCol", ascending: false),
+            SortDescriptor(keyPath: "intCol", ascending: false)])
         XCTAssertEqual(2.22, sorted[0].doubleCol)
         XCTAssertEqual(3, sorted[0].intCol)
         XCTAssertEqual(2.22, sorted[1].doubleCol)
         XCTAssertEqual(2, sorted[1].intCol)
         XCTAssertEqual(1.11, sorted[2].doubleCol)
 
-        assertThrows(collection.sorted([SortDescriptor(property: "noSuchCol")]), named: "Invalid sort property")
+        assertThrows(collection.sorted([SortDescriptor(keyPath: "noSuchCol")]), named: "Invalid property name")
     }
 
     func testMin() {
@@ -1841,9 +1849,9 @@ class ListUnmanagedRealmCollectionTypeTests: ListRealmCollectionTypeTests {
 
     override func testSortWithDescriptor() {
         let collection = getAggregateableCollection()
-        assertThrows(collection.sorted([SortDescriptor(property: "intCol", ascending: true)]))
-        assertThrows(collection.sorted([SortDescriptor(property: "doubleCol", ascending: false),
-            SortDescriptor(property: "intCol", ascending: false)]))
+        assertThrows(collection.sorted([SortDescriptor(keyPath: "intCol", ascending: true)]))
+        assertThrows(collection.sorted([SortDescriptor(keyPath: "doubleCol", ascending: false),
+            SortDescriptor(keyPath: "intCol", ascending: false)]))
     }
 
     override func testFastEnumerationWithMutation() {